more refactor
* split some tests from test_base.py
* dont approve sts version gates for now
